+++ lib/Datasource/File.php
@@ -96,7 +96,8 @@ class File implements IDatasource
         $delimiter = $this->detectDelimiter($rows[0]);
 
         // the first row will define the column headers, even if it is not a real header
-        $header = str_getcsv($rows[0], $delimiter);
+        // trim removes any leading or ending spaces
+        $header = array_map('trim', str_getcsv($rows[0], $delimiter));
 
         // if the data has a real header, remove the first row
         if (!isset($option['hasHeader']) or $option['hasHeader'] !== 'false') {
@@ -108,11 +109,11 @@ class File implements IDatasource
             // if only a subset of columns or fixed column values are set, they are replaced here
             $header = $this->minimizeRow($selectedColumns, $header);
             foreach ($rows as $row) {
-                $data[] = $this->minimizeRow($selectedColumns, str_getcsv($row, $delimiter));
+                $data[] = $this->minimizeRow($selectedColumns, array_map('trim', str_getcsv($row, $delimiter)));
             }
         } else {
             foreach ($rows as $row) {
-                $data[] = str_getcsv($row, $delimiter);
+                $data[] = array_map('trim', str_getcsv($row, $delimiter));
             }
         }
         unset($rows);
@@ -145,7 +146,6 @@ class File implements IDatasource
         $data_2 = array();
         $delimiter = $delimiters[0];
         foreach ($delimiters as $d) {
-            //$firstRow = str_getcsv($data, "\n")[0];
             $data_1 = str_getcsv($data, $d);
             if (sizeof($data_1) > sizeof($data_2)) {
                 $delimiter = $d;